% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{wrapper}
\alias{wrapper}
\alias{r6_class}
\title{Wrappers}
\usage{
wrapper(x, y)
}
\arguments{
\item{x}{x param}

\item{y}{y param}
}
\description{
Description of wrapper
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{result}}{The result}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-r6_class-new}{\code{r6_class$new()}}
\item \href{#method-r6_class-get_result}{\code{r6_class$get_result()}}
\item \href{#method-r6_class-clone}{\code{r6_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r6_class-new"></a>}}
\if{latex}{\out{\hypertarget{method-r6_class-new}{}}}
\subsection{Method \code{new()}}{
makes object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r6_class$new(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r6_class-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-r6_class-get_result}{}}}
\subsection{Method \code{get_result()}}{
get result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r6_class$get_result(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r6_class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-r6_class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r6_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
