% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rschtasks-package.R
\docType{package}
\name{rschtasks-package}
\alias{rschtasks}
\alias{rschtasks-package}
\title{rschtasks: Interface for schtasks}
\description{
An R interface for scheduling task on Windows.
}
\author{
\strong{Maintainer}: Jordan Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
